Accurate prices for etfs and UK equities

+++ sql/schema/002_add_open_price_to_positions.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 alter table positions
-add column previous_close_price double precision;
+add column previous_close_price bigint;
 
 -- +goose Down
 alter table positions
